move some files related to plugin API implementation to 'plugin' folder
